Mention Linux packages on install page (#9314)
* Update install docs to mention Linux packages

We now build packages for Debian, Ubuntu, CentOS, etc. This removes language
about "we have no plans to build packages" and adds links to step by step guides
for adding a GPG key and the official repository.

* Fix URL to Learn Vault install page

A Linux section previously existed but now it is in the general install section.

* Fix Markdown for multi-step compile from source

The steps were previously marked up as an ordered list but the numbers didn't
display correctly. This outdents the code so it's a series of paragraphs instead
of an ordered list.